k230-inference/.claude/worktrees/romantic-satoshi-501183/k230-training/workspace/requirements-train.txt
# Training Environment Requirements
# For use with the 'train' Docker service

# Deep Learning Framework
torch==2.1.0
torchvision==0.16.0

# YOLOv8 and Dependencies
ultralytics==8.1.0

# ONNX Export
onnx==1.15.0
onnxsim==0.4.35
onnxruntime-gpu==1.16.3

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0

# Data Processing
numpy==1.24.3
pandas==2.1.4

# Configuration and Utilities
pyyaml==6.0.1
tqdm==4.66.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Experiment Tracking
tensorboard==2.15.1

# Additional utilities
scipy==1.11.4
scikit-learn==1.3.2
